Move LocalSessionStorage and WebViewData tests to privacy.
Bundling tests by function will help scanning what we actually test.
